requirements.txt
fastapi
uvicorn
pydantic>=2.0
groq
python-dotenv
pytest
pytest-asyncio